Tighten browser new-tab request boundaries
